Cleanup: remove ignored flags from poll events bitmasks
The POLLHUP and POLLERR flags are only valid in 'revents', they are
always checked regardless of the fact they were set in 'events' or not.

As such remove those flags from all poll events to reduce possible
confusion as to which flags can be returned by poll.

This is in preparation of a patch for FreeBSD to work around a bug where
polling a pipe fd with an empty 'events' will hang.

Change-Id: Id22c78c38257d96dfc47e1337795f13c70dd5f91
Signed-off-by: Michael Jeanson <mjeanson@efficios.com>