wip: POLLIN
The change in the agent application handling solves the following
problem:

The agent thread assumes that any POLLIN event occurs on the
registration socket (i.e. a new app is registering). This is true on
Linux, but we receive "POLLIN | POLLHUP" when an application dies on
FreeBSD. This ends up confusing the agent thread.

The patch ensures that the agent thread explicitly checks on which 'fd'
events occur and reacts accordingly.

Change-Id: Ie4de3354491c17c346258022ec9584d79e9f3ff2