tests: tap-driver.sh: disable input buffering with mawk
Contrary to other AWK implementations, mawk buffers its input by default
which results on the TAP driver receiving the whole output of a test
only as the process exits. This can be problematic for inactivity
timeout checks based on the output as well as inconvenient when
interactively following a test run.

Mawk supports a non-standard 'interactive' mode which disables input and
output buffering. To work around this behavior, check if the current AWK
binary supports this option and enable it.

Change-Id: Ida10dfec8a552012fa411c623acc438debdaca1d
Signed-off-by: Michael Jeanson <mjeanson@efficios.com>